+++ portals/publisher/src/main/webapp/source/src/app/components/Shared/PoliciesUI/PolicyViewForm.tsx
